Added docker tests and solved network proxy issues
