+++ node_engine/NetManager/proxy/proxyCache.go
@@ -31,7 +31,7 @@ type ProxyCache struct {
 func NewProxyCache() ProxyCache {
 	return ProxyCache{
 		cache:                 make(map[int]ConversionList),
-		conversionListMaxSize: 30,
+		conversionListMaxSize: 10,
 		rwlock:                sync.RWMutex{},
 	}
 	//TODO: Start cleanup procedure each X seconds
@@ -42,11 +42,11 @@ func (cache *ProxyCache) RetrieveByServiceIP(srcIP net.IP, srcport int, dstServi
 	cache.rwlock.Lock()
 	defer cache.rwlock.Unlock()
 
-	elem, exist := cache.cache[dstport]
+	elem, exist := cache.cache[srcport]
 	if exist {
 		elem.lastUsed = time.Now().Unix()
 		for _, entry := range elem.conversionList {
-			if entry.srcport == srcport && entry.dstServiceIp.Equal(dstServiceIp) && entry.srcip.Equal(srcIP) {
+			if entry.dstport == dstport && entry.dstServiceIp.Equal(dstServiceIp) && entry.srcip.Equal(srcIP) {
 				return entry, true
 			}
 		}
@@ -59,11 +59,11 @@ func (cache *ProxyCache) RetrieveByInstanceIp(srcip net.IP, srcport int, dstport
 	cache.rwlock.Lock()
 	defer cache.rwlock.Unlock()
 
-	elem, exist := cache.cache[dstport]
+	elem, exist := cache.cache[srcport]
 	if exist {
 		elem.lastUsed = time.Now().Unix()
 		for _, entry := range elem.conversionList {
-			if entry.srcport == srcport && entry.srcip.Equal(srcip) {
+			if entry.dstport == dstport && entry.srcip.Equal(srcip) {
 				return entry, true
 			}
 		}
@@ -76,11 +76,11 @@ func (cache *ProxyCache) Add(entry ConversionEntry) {
 	cache.rwlock.Lock()
 	defer cache.rwlock.Unlock()
 
-	_, exist := cache.cache[entry.dstport]
+	_, exist := cache.cache[entry.srcport]
 	if exist {
 		cache.addExisting(entry)
 	} else {
-		cache.cache[entry.dstport] = ConversionList{
+		cache.cache[entry.srcport] = ConversionList{
 			nextEntry:      0,
 			lastUsed:       time.Now().Unix(),
 			conversionList: make([]ConversionEntry, cache.conversionListMaxSize),
@@ -90,13 +90,13 @@ func (cache *ProxyCache) Add(entry ConversionEntry) {
 }
 
 func (cache *ProxyCache) addExisting(entry ConversionEntry) {
-	elem, _ := cache.cache[entry.dstport]
+	elem, _ := cache.cache[entry.srcport]
 	elem.lastUsed = time.Now().Unix()
 	alreadyExist := false
 	alreadyExistPosition := 0
 	//check if used port is already in proxycache
 	for i, elementry := range elem.conversionList {
-		if elementry.srcport == entry.srcport {
+		if elementry.dstport == entry.dstport {
 			alreadyExistPosition = i
 			alreadyExist = true
 			break
@@ -112,5 +112,3 @@ func (cache *ProxyCache) addExisting(entry ConversionEntry) {
 		elem.nextEntry = (elem.nextEntry + 1) % cache.conversionListMaxSize
 	}
 }
-
-//TODO add tests
